Materials/requirements.txt
apache-airflow-providers-cncf-kubernetes=7.9.0